src/handlers/reminders.py
from telegram.ext import CommandHandler
import asyncio

async def remind(update: Update, context: ContextTypes.DEFAULT_TYPE):
    try:
        delay, task = parse_reminder(context.args)  # Implement parsing logic
        await asyncio.sleep(delay)
        await update.message.reply_text(f"🔔 Reminder: {task}")
    except Exception as e:
        await update.message.reply_text("⚠️ Usage: /remindme <time> <task>")

def setup_reminders(application):
    application.add_handler(CommandHandler("remindme", remind))