requirements.txt
python-telegram-bot==20.3
google-generativeai>=0.3.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
apscheduler>=3.10.0